Add files via upload
In this commit, I just added the footer.  I added the nav bar and inside the nav bar will be links to the marvel website, my Instagram Website and my Github link.

I want to make sure when reviewers view this website, they are able to go to the main page and meet me as well.